Create sm-samba-detect-group-changes.sh
Samba has no built in capability to check for group changes.  This takes care of that by doing an ldapsearch on groups.  This creates a file with a bytesize.  The cron runs this script every X minutes to see if ldapsearch results in a different bytesize amount than the last one.  If so, something in groups changed, run your favorite script.