Update sm_nc_add_auth_group_privilege
The purpose of this script is to:
NextCloud does not appear to have a built in occ command for adding Admin Right Privileges aka authorized group privileges as described here: 
 https://docs.nextcloud.com/server/latest/admin_manual/configuration_server/admin_delegation_configuration.html

Until they do, this script provides a way for you to do it directly on your NC database.

We have been using it and found it to cause no side effects and implement the solution advertised.  That said, this script is given freely and as-as where is.